requirements.txt
openai
chromadb
python-dotenv
streamlit
pandas